+++ stroom-query-testing/src/integration-test/java/stroom/query/testing/jooq/TestJooqQueryResourceIT.java
@@ -9,18 +9,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import stroom.datasource.api.v2.DataSource;
 import stroom.datasource.api.v2.DataSourceField;
-import stroom.query.api.v2.DocRef;
-import stroom.query.api.v2.ExpressionOperator;
-import stroom.query.api.v2.ExpressionTerm;
-import stroom.query.api.v2.Field;
-import stroom.query.api.v2.FlatResult;
-import stroom.query.api.v2.OffsetRange;
-import stroom.query.api.v2.Query;
-import stroom.query.api.v2.Result;
-import stroom.query.api.v2.ResultRequest;
-import stroom.query.api.v2.SearchRequest;
-import stroom.query.api.v2.SearchResponse;
-import stroom.query.api.v2.TableSettings;
+import stroom.query.api.v2.*;
 import stroom.query.audit.authorisation.DocumentPermission;
 import stroom.query.audit.model.DocRefEntity;
 import stroom.query.audit.rest.AuditedDocRefResourceImpl;
@@ -28,11 +17,7 @@ import stroom.query.testing.DropwizardAppWithClientsRule;
 import stroom.query.testing.QueryResourceIT;
 import stroom.query.testing.StroomAuthenticationRule;
 import stroom.query.testing.data.CreateTestDataClient;
-import stroom.query.testing.jooq.app.CreateTestDataJooqImpl;
-import stroom.query.testing.jooq.app.JooqApp;
-import stroom.query.testing.jooq.app.JooqConfig;
-import stroom.query.testing.jooq.app.TestDocRefJooqEntity;
-import stroom.query.testing.jooq.app.TestQueryableJooqEntity;
+import stroom.query.testing.jooq.app.*;
 
 import javax.ws.rs.core.Response;
 import java.util.Collection;
@@ -43,9 +28,7 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import static io.dropwizard.testing.ResourceHelpers.resourceFilePath;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.junit.Assert.*;
 
 public class TestJooqQueryResourceIT extends QueryResourceIT<TestDocRefJooqEntity, JooqConfig> {
 
@@ -57,7 +40,7 @@ public class TestJooqQueryResourceIT extends QueryResourceIT<TestDocRefJooqEntit
 
     @ClassRule
     public static StroomAuthenticationRule authRule =
-            new StroomAuthenticationRule(WireMockConfiguration.options().port(10080), TestDocRefJooqEntity.TYPE);
+            new StroomAuthenticationRule(WireMockConfiguration.options().port(10080));
 
     private final CreateTestDataClient testDataClient;
 
@@ -82,8 +65,14 @@ public class TestJooqQueryResourceIT extends QueryResourceIT<TestDocRefJooqEntit
                 .name(UUID.randomUUID().toString())
                 .type(TestDocRefJooqEntity.TYPE)
                 .build();
-        authRule.giveFolderCreatePermission(authRule.adminUser(), parentFolderUuid);
-        authRule.giveDocumentPermission(authRule.adminUser(), testDataDocRef.getUuid(), DocumentPermission.READ);
+        authRule.permitAdminUser()
+                .createInFolder(parentFolderUuid)
+                .docRefType(TestDocRefJooqEntity.TYPE)
+                .done();
+        authRule.permitAdminUser()
+                .docRef(testDataDocRef)
+                .permission(DocumentPermission.READ)
+                .done();
 
         final Response createDocumentResponse = docRefClient.createDocument(authRule.adminUser(), testDataDocRef.getUuid(), testDataDocRef.getName(), parentFolderUuid);
         assertEquals(HttpStatus.OK_200, createDocumentResponse.getStatus());